+++ packages/api/tests/lib/writeRecords/test-write-granules.js
@@ -940,7 +940,7 @@ test.serial('writeGranulesFromMessage() given an empty array as a files key will
   });
 
   // Files were removed from all datastores
-  t.is(apiGranule.files, undefined);
+  t.deepEqual(apiGranule.files, []);
   t.deepEqual(dynamoGranule.files, []);
   t.deepEqual(esGranule.files, []);
 });
@@ -2121,7 +2121,7 @@ test.serial('writeGranuleFromApi() given an empty array as a files key will remo
   });
 
   // Files were removed from all datastores
-  t.is(apiGranule.files, undefined);
+  t.deepEqual(apiGranule.files, []);
   t.deepEqual(dynamoGranule.files, []);
   t.deepEqual(esGranule.files, []);
 });